+++ src/router/index.ts
@@ -4,7 +4,7 @@ import StudentHomeView from "../views/student/Home.vue";
 import LecturerHomeView from "../views/lecturer/Home.vue";
 import AdminHomeView from "../views/admin/Home.vue";
 import CourseFormSuspenseWrapper from "../views/lecturer/CourseFormSuspenseWrapper.vue";
-import AccountFormSuspenseWrapper from "../views/admin/AccountFormSuspenseWrapper.vue";
+import AccountFormSuspenseWrapper from "../views/admin/EditCreateAccountForm.vue";
 import Redirect from "../views/errors/403.vue";
 import ProfileWrapper from "../components/profile/Wrapper.vue";
 import Settings from "../views/common/Settings.vue";
