+++ src/views/admin/EditCreateAccountForm.vue
@@ -1,5 +1,5 @@
 <template>
-    <div class="w-full h-screen mx-auto mt-8 bg-gray-300 lg:mt-20">
+    <div v-if="!busy" class="w-full h-screen mx-auto mt-8 bg-gray-300 lg:mt-20">
         <button id="navigateBack" class="flex items-center mb-4 navigation-link" @click="back()">
             <i class="text-xl fas fa-chevron-left"></i>
             <span class="ml-1 text-sm font-bold">Back</span>
@@ -123,14 +123,18 @@
                 </div>
             </section>
             <delete-account-modal ref="deleteModal" />
+            <unsaved-changes-modal ref="unsavedChangesModal" />
         </div>
+    </div>
+    <div v-else>
+        <loading-component />
     </div>
 </template>
 
 <script lang="ts">
     import Router from "@/router/";
     import { Role } from "@/entities/Role";
-    import { ref, reactive, computed } from "vue";
+    import { ref, reactive, computed, onBeforeMount } from "vue";
     import { FieldOfStudy } from "@/api/api_models/user_management/FieldOfStudy";
     import UserManagement from "@/api/UserManagement";
     import StudentEntity from "@/entities/StudentEntity";
@@ -152,6 +156,9 @@
     import PersonalInformationSection from "@/components/account/edit/PersonalInformationSection.vue";
     import LecturerInformationSection from "@/components/account/edit/LecturerInformationSection.vue";
     import StudentInformationSection from "@/components/account/edit/StudentInformationSection.vue";
+    import LoadingComponent from "../../components/loading/Spinner.vue";
+    import { checkPrivilege } from "@/use/PermissionHelper";
+    import UnsavedChangesModal from "@/components/modals/UnsavedChangesModal.vue";
 
     export default {
         name: "AdminCreateAccountForm",
@@ -162,6 +169,46 @@
             PersonalInformationSection,
             LecturerInformationSection,
             StudentInformationSection,
+            UnsavedChangesModal,
+            LoadingComponent,
+        },
+        async beforeRouteEnter(_to: any, _from: any, next: any) {
+            const response = await checkPrivilege(Role.ADMIN);
+
+            if (response.allowed) {
+                return next();
+            }
+            if (!response.authenticated) {
+                return next("/login");
+            }
+
+            return next("/redirect");
+        },
+
+        async beforeRouteLeave(to: any, from: any, next: any) {
+            if (this.success) {
+                return next();
+            }
+            if (this.hasInput) {
+                const modal = this.unsavedChangesModal;
+                let action = modal.action;
+                const response = await modal.show();
+                switch (response) {
+                    case action.CANCEL: {
+                        next(false);
+                        break;
+                    }
+                    case action.CONFIRM: {
+                        next(true);
+                        break;
+                    }
+                    default: {
+                        next(true);
+                    }
+                }
+            } else {
+                next(true);
+            }
         },
         props: {
             editMode: {
@@ -170,7 +217,8 @@
             },
         },
         emits: ["update:has-input", "update:success"],
-        async setup(props: any, { emit }: any) {
+        setup(props: any, { emit }: any) {
+            let busy = ref(false);
             let account = reactive({
                 authUser: new Account(),
                 user: new UserEntity(),
@@ -189,6 +237,7 @@
             let title = props.editMode ? "Account Editing" : "Account Creation";
             let success = ref(false);
             let deleteModal = ref();
+            let unsavedChangesModal = ref();
 
             const errorBag: ErrorBag = reactive(new ErrorBag());
 
@@ -200,7 +249,14 @@
                 return account.user.role === Role.STUDENT;
             });
 
-            if (props.editMode) {
+            onBeforeMount(() => {
+                if (props.editMode) {
+                    getUser();
+                }
+            });
+
+            async function getUser() {
+                busy.value = true;
                 const userManagement: UserManagement = new UserManagement();
 
                 const response = await userManagement.getSpecificUser(Router.currentRoute.value.params.username as string);
@@ -224,6 +280,7 @@
                         initialAccount.admin = JSON.parse(JSON.stringify(account.admin));
                     }
                 }
+                busy.value = false;
             }
 
             function updatePicture() {}
@@ -382,6 +439,7 @@
             }
 
             return {
+                busy,
                 title,
                 account,
                 success,
@@ -395,6 +453,7 @@
                 deleteAccount,
                 confirmDeleteAccount,
                 deleteModal,
+                unsavedChangesModal,
                 errorBag: errorBag,
             };
         },
